dom/security/test/csp/test_nonce_redirects.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Bug 1469150:Scripts with valid nonce get blocked if URL redirects</title>
  <!-- Including SimpleTest.js so we can use waitForExplicitFinish !-->
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<iframe style="width:100%;" id="testframe"></iframe>

<script class="testbody" type="text/javascript">

/* Description of the test:
 * We load a script with a matching nonce, which redirects
 * and we make sure that script is allowed.
 */

SimpleTest.waitForExplicitFinish();

function finishTest() {
  window.removeEventListener("message", receiveMessage);
  SimpleTest.finish();
}

function checkResults(aResult) {

  if (aResult === "script-loaded") {
    ok(true, "expected result: script loaded");
  }
  else {
    ok(false, "unexpected result: script blocked");
  }
  finishTest();
}

window.addEventListener("message", receiveMessage);
function receiveMessage(event) {
  checkResults(event.data.result);
}

document.getElementById("testframe").src = "file_nonce_redirects.html";

</script>
</body>
</html>
